Fix the saving of individual files
This commit fixes the saving of individual files by using
fp.file.parent instead of fp.displayDirectory to determine the parent
directory of the file to be saved that was chosen by the user.

According to [1], nsIFilePicker's "displayDirectory" attribute refers
to "the directory that the file picker should initially display", and
that it "should be set before calling open() or show()", whereas the
"file" attribute is file/directory selected by the user as a result of
the file picker dialog.

In addition, according to [2], nsIFilePicker's "show" method has been
obsolete since Gecko 57.0, so this commit removes the use of the show
method.

Fixes: #16

[1] https://developer.mozilla.org/en-US/docs/Mozilla/Tech/XPCOM/Reference/Interface/nsIFilePicker#Attributes
[2] https://developer.mozilla.org/en-US/docs/Mozilla/Tech/XPCOM/Reference/Interface/nsIFilePicker#show()_2